separated match result view and view model
